+++ src/keyboard.c
@@ -114,7 +114,7 @@ void putinbuffer(unsigned char ascii){
 		//__write(STD_OUT,keyboard_buffer,last-first);
 }
 
-void forBuffer(unsigned char scancode, short unsigned int gs, short unsigned int fs, short unsigned int es, short unsigned int ds, short unsigned int ss, int edi, int esi, int ebx, int edx, int ecx, int esp, short unsigned int cs, int eip,int eax, int flags,int ebp){
+void forBuffer(unsigned char scancode, short unsigned int gs, short unsigned int fs, short unsigned int es, short unsigned int ds, short unsigned int ss, int edi, int esi, int ebx, int edx, int ecx, int esp, short unsigned int cs, int eip,int eax, unsigned short int flags,int ebp){
     unsigned char ascii=ZERO;
    	int specialkeynum=-1; 
    	int specialindex;
